started removing builder pattern
